+++ src/hooks/useProducts.ts
@@ -12,27 +12,48 @@ export default function useProducts() {
         language,
     }: {
         name: string;
-        picture: string;
+        picture: File;
         description: string;
         language: string;
     }) => {
-        const res = await fetch("/api/products", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify({
-                name, 
-                picture,
-                description,
-                language,
-            }),
-        });
-        if (!res.ok) {
-            const body = await res.json();
-            throw new Error(body.error);
+        try {
+            // upload image to cloud and get the url
+            const formData = new FormData();
+            formData.append("file", picture);
+
+            const imageRes = await fetch("/api/image", {
+                method: "POST",
+                body: formData,
+            });
+            if (!imageRes.ok) {
+                const error = await imageRes.json();
+                alert("Error uploading image");
+                throw new Error(error);
+            }
+            const imageUrl = await imageRes.json();
+            console.log("imageUrl: ", imageUrl);
+
+            // upload the image url and data to the database
+            const res = await fetch("/api/products", {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify({
+                    name, 
+                    picture,
+                    description,
+                    language,
+                }),
+            });
+            if (!res.ok) {
+                const body = await res.json();
+                throw new Error(body.error);
+            }
+            router.refresh();
+        } catch (error) {
+            console.error("error: ", error);
         }
-        router.refresh();
     }
 
     //GET
@@ -60,9 +81,57 @@ export default function useProducts() {
     }: {
         id: number;
         name: string;
-        picture: string;
+        picture: File;
         description: string;
     }) => {
+        // get the image url from the database
+        const resGet = await fetch("/api/products", {
+            method: "GET",
+            headers: {
+                "Content-Type": "application/json",
+            },
+        });
+        if (!resGet.ok) {
+            const body = await resGet.json();
+            throw new Error(body.error);
+        }
+        const productsList = await resGet.json();
+        const products = productsList.products;
+        const target = products.find((product: any) => product.id === id);
+        const url = target.picture;
+
+        // delete image from cloud
+        const resDelete = await fetch("/api/image", {
+            method: "DELETE",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            body: JSON.stringify({
+                url,
+            }),
+        });
+        if (!resDelete.ok) {
+            const body = await resDelete.json();
+            throw new Error(body.error);
+        }
+
+        // upload the new image to cloud
+        const formData = new FormData();
+        formData.append("file", picture);
+
+        const imageRes = await fetch("/api/image", {
+            method: "POST",
+            body: formData,
+        });
+        if (!imageRes.ok) {
+            const error = await imageRes.json();
+            alert("Error uploading image");
+            throw new Error(error);
+        }
+        const imageUrl = await imageRes.json();
+        console.log("imageUrl: ", imageUrl);
+
+        // update the object in the database
         const res = await fetch('/api/products', {
             method: "PUT",
             headers: {
@@ -84,6 +153,38 @@ export default function useProducts() {
 
     //DELETE
     const deleteProducts = async (id: number) => {
+        // get the image url from the database
+        const resGet = await fetch("/api/products", {
+            method: "GET",
+            headers: {
+                "Content-Type": "application/json",
+            },
+        });
+        if (!resGet.ok) {
+            const body = await resGet.json();
+            throw new Error(body.error);
+        }
+        const productsList = await resGet.json();
+        const products = productsList.products;
+        const target = products.find((product: any) => product.id === id);
+        const url = target.picture;
+
+        // delete image from cloud
+        const resDelete = await fetch("/api/image", {
+            method: "DELETE",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            body: JSON.stringify({
+                url,
+            }),
+        });
+        if (!resDelete.ok) {
+            const body = await resDelete.json();
+            throw new Error(body.error);
+        }
+
+        // delete the object from the database
         const res = await fetch('/api/products', {
             method: "DELETE",
             headers: {
